This update fixes pathing to remove SRCMLBINDIR. Some fixes to the files used to create the .dll (now in their own separate namespace, changed name of library) SrcML files are now copied from external/srcml1.0/bin to build/debug Removed old srcMLCppApi project
