+++ SrcML.NET.sln
@@ -63,7 +63,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ABB.SrcML.Tools.MethodCallS
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ABB.SrcML.Data.Test", "ABB.SrcML.Data.Test\ABB.SrcML.Data.Test.csproj", "{2D910B34-7FEE-44DC-B79F-00DA2A914295}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SrcMLCppAPI", "SrcMLCppAPI\SrcMLCppAPI.vcxproj", "{577E8416-DE0E-4675-A31F-49FFDF49F4D3}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "LibSrcMLWrapper", "LibSrcMLWrapper\LibSrcMLWrapper.vcxproj", "{5A139FFE-60CB-4C14-B565-73009393B47A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -306,20 +306,20 @@ Global
 		{2D910B34-7FEE-44DC-B79F-00DA2A914295}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{2D910B34-7FEE-44DC-B79F-00DA2A914295}.Release|Win32.ActiveCfg = Release|Any CPU
 		{2D910B34-7FEE-44DC-B79F-00DA2A914295}.Release|x86.ActiveCfg = Release|Any CPU
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|Win32.ActiveCfg = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|Win32.Build.0 = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|x86.ActiveCfg = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Debug|x86.Build.0 = Debug|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|Any CPU.ActiveCfg = Release|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|Win32.ActiveCfg = Release|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|Win32.Build.0 = Release|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|x86.ActiveCfg = Release|Win32
-		{577E8416-DE0E-4675-A31F-49FFDF49F4D3}.Release|x86.Build.0 = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|Any CPU.ActiveCfg = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|Mixed Platforms.Build.0 = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|Win32.ActiveCfg = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|Win32.Build.0 = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|x86.ActiveCfg = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Debug|x86.Build.0 = Debug|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|Any CPU.ActiveCfg = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|Mixed Platforms.ActiveCfg = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|Mixed Platforms.Build.0 = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|Win32.ActiveCfg = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|Win32.Build.0 = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|x86.ActiveCfg = Release|Win32
+		{5A139FFE-60CB-4C14-B565-73009393B47A}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
